requirements.txt
asgiref==3.3.4
beautfulsoup4==4.5.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
djangorestframework==3.12.4
idna==2.10
lxml==4.6.3
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
